+++ fcg/microscopy/__init__.py
@@ -1,1 +1,2 @@
+from ._tract import Tracts, load_tracts
 from ._iscat_movie_sim import MovieAcquisitionSimulator
