+++ scripts/sim_microscope_3d_acquisition.py
@@ -7,7 +7,7 @@ import typing
 import colorama
 
 import fcg.app
-from fcg.microscopy import MovieAcquisitionSimulator
+from fcg.microscopy import load_tracts, MovieAcquisitionSimulator
 
 _DEFAULT_OUT_PATH: typing.Final[pathlib.Path] = pathlib.Path("out.tiff")
 
@@ -45,22 +45,32 @@ class SimulateMicroscope3dAcquisition(fcg.app.App):
             "--contrast",
             default=5,
             type=float,
-            help="Contrast between the simulated particle and the background (Contrast = particle "
+            help=" Contrast between the simulated particle and the background (Contrast = particle "
             "intensity - background intensity) %(default)s",
         )
         parser.add_argument("--background_intensity", default=0.3, type=float, help="Background intensity %(default)s")
         parser.add_argument(
             "--gaussian_noise_variance",
             default=0.15,
-            type=float,
+            type=typing.Union[float, None],
             help="Gaussian noise variance (default=%(default)s)",
         )
         parser.add_argument("--poisson_noise", action="store_true", help="Add poisson noise after PSF convolution")
 
         args = parser.parse_args()
 
+        print("Loading the tracts ... ", end="")
+        try:
+            fib: pathlib.Path = args.fib
+            fib = fib.resolve(strict=True)
+            tracts = load_tracts(fib)
+            print(colorama.Style.BRIGHT + colorama.Fore.GREEN + "succeeded")
+        except Exception as exception:
+            print(colorama.Style.BRIGHT + colorama.Fore.RED + "failed")
+            raise exception
+
         simulator = MovieAcquisitionSimulator(
-            args.tracks,
+            tracts=tracts,
             resolution=args.resolution,
             dt=args.time_resolution,
             contrast=args.contrast,
@@ -68,17 +78,6 @@ class SimulateMicroscope3dAcquisition(fcg.app.App):
             noise_gaussian=args.gaussian_noise_variance,
             noise_poisson=args.poisson_noise,
         )
-
-        print("Loading the tracts ... ", end="")
-        try:
-            fib: pathlib.Path = args.fib
-            fib = fib.resolve(strict=True)
-
-            simulator.load_tracts(fib)
-            print(colorama.Style.BRIGHT + colorama.Fore.GREEN + "succeeded")
-        except Exception as exception:
-            print(colorama.Style.BRIGHT + colorama.Fore.RED + "failed")
-            raise exception
 
         print("Loading the PSF ... ", end="")
         try:
